Pass opaque types by UniquePtr where value needed.
When existing C++ APIs demand that a non-POD type be passed
by value (e.g. a std::string or a struct which is not trivial)
we would like those APIs to be callable from Rust, but by
passing a UniquePtr<T> instead.

That's what this PR does.

It relies on cxx PR #218 so we can't merge it yet.