Adding __gnu to blocklist
On my Linux box, without this we ended up trying to make lots
of cxx entries for types beginning with __gnu. Like the previous
entry for _mbstate_t_, this should not theoretically be necessary
as bindgen should not be outputting these entries if they're not used
by something else in the allowlist. But it does. And I imagine that's
because this sort of scary template dependency chain is hard to trace
in practice, so blocklist for now.