Abandon temp dirs in build system.
This was a bit of a hack anyway, and has ceased to work now
we are writing an include file which needs to be made available
at macro expansion time.

This commit instead writes the generated C++/header files to the
cargo out directory. It does this in a blugeoningly naive way
compared to cxx, so undoubtedly there will be dozens of follow
up commits.

In particular it makes no effort to clear out older irrelevant
files.