First part of namespace work.
This change asks bindgen to generate information about C++ namespaces.
bindgen does this by outputting a hierarchic structure of 'mods'. This change
contains all the changes required to autocxx to be able to handle the bindgen
input when it's generated in that mode.

It doesn't yet actually pass that namespace information across to cxx,
and therefore this is useless at actually working with namespaces. That
will come in the future.