Change macro syntax.
This is a BREAKING CHANGE which switches from include_cxx!
with a certain syntax to include_cpp! with a different syntax.

The main goal here is to be able to attach rustdoc to the
directives that are allowed within the macro. Thanks to dtolnay@
for the macro_rules magic which allows this.

A future commit will actually add such docs.