Static methods, and deconflict names.
This is a significant set of interrelated changes.

The main goals are:
* Support static member functions
* Handle conflicting function and method names correctly.

Identifying static member functions is a bit awkward as it
requires us to hold off on converting a given ForeignFn until
we've looked in subsequent impl blocks that are generated by
bindgen. Otherwise, this is relatively straightforward.

Deconflicting names is more complex. See the comment in
bridge_name_tracker.rs.