Require 'unsafe' keyword.
This matches the current policy of the cxx crate which requires
cxx::bridge blocks to be marked as 'unsafe', or alternatively
all individual functions themselves will be marked as 'unsafe' and
thus need to be called in 'unsafe' blocks.

With this commit, autocxx only allows the former policy: the
include_cpp! block must be marked unsafe.

This is not entirely satisfactory, as pragmatically some C++
FFI calls are obviously _more_ unsafe than others, and from
a code review perspective it would be desirable for them to be
marked with extra 'unsafe'ness. It's likely that a more nuanced
policy will be adopted in future.