Move POD checking to an analysis phase.
Previously, we walked the bindgen output twice (in effect) -
once to identify POD-safe types, and then again to assemble
our set of APIs. Both of these phases were done in
the parse sub-module.

Now, we assemble the set of APIs once, and apply a
subsequent analysis phase to that set of APIs to decorate
them with information on what types are POD-safe.

This is a bit more code at the moment but is architecturally
cleaner. As we further break down the parse sub-module into
further analysis phases, it should result in a cleaner
design at the end of it all.