Create TypeConverter in parse phase.
TypeConverter remains an object which spans multiple phases
of parsing, analysis, and codegen. We should eventually zap it
and store all its information within the API list - but we're
not at that point yet.

(The primary information that it stores is the typedef mappings.)

However, it's an output of the parsing phase, so create it as
such.