Fix issues 217, 222.
This fixes two issues where autocxx-bindgen was being too aggressive
in trying to denote whether a given type was a reference or pointer.
We'll now take such particular cases to be pointers even if they're
references, which is the conservative and safe choice but causes more
friction for callers.

Follow up work will be required when we properly implement
the corresponding features:
https://github.com/google/autocxx/issues/227
https://github.com/google/autocxx/issues/228