Add error symbols into ourput mod.
There are some APIs which autocxx can't generate because they are
unsupported C++ features.

Previously, the codegen phase reported an error to stderr.
Now, it does that, and also adds a ZST into the generated
code with a doc comment indicating the trouble. The idea is that
those using rust-analyzer or some other autocompletion system
will see an error message indicating the problem.

At present, this change does not attempt to add impl blocks
for problematic methods. A future change may do that.