Allow passing arbitrary extra Clang arguments.
Some build environments need this capability, for example to set up a
different `-sysroot` or to specify `-iquote` and `-isystem` directories.
This feature also makes it possible to specify a different C++ version
-- see https://github.com/google/autocxx/issues/209.

This change makes the recently added feature for specifying preprocessor
defines obsolete, so it removes that feature.