Support competing wrappers
The primary difficulty here is that autocxx has some built-in
"utility" functions, specifically make_string. To avoid ODR violations
we need to prefix those symbols with the mod name. To do that,
we need to squash the TypeConfig into the overall IncludeCppConfig
since the allowlist will now depend upon the mod name.