Inline all autocxx-generated C++ functions.
autocxx generates two lumps of C++ code:
* code generated by cxx (both .h and .cpp)
* code generated by autocxx (previously, both .h and .cpp)

In fact all functions within the second area of code are
only ever called from the cxx-generated .cpp file.
This change ensures all our functions (in the second generated
section of code) are inlined, to avoid the waste of an extra
C++ compiler invocation, and to encourage inlining even in
build systems where link-time optimization may not be enabled.

This first change still creates a blank .cpp file for the
second set of bindings. Subsequent commits will eliminate that.