Remove `bindgen_` attributes from output.
These attributes don't have a corresponding macro defined anywhere, so
they will cause compilation errors if we leave them in.

This bug was exposed by a recent change in autocxx-bindgen that added
`bindgen_original_name` attributes for nested types; this is why
autocxx-bindgen was temporarily pinned to version 0.58.3, which does not
contain those changes.

Because the bug is now fixed, this change also updates autocxx-bindgen
to 0.58.4.