Preprocess mode: Check env vars `CLANG_PATH` and `CXX` for path to Clang.
This is to support environments that need to use a different `clang++`
than the one on the path. `CLANG_PATH` is the variable used by the
clang-sys crate; in most cases where a custom Clang is needed, people
will already be specifying this, and things should just work.

Also improve the error reporting when preprocessing goes wrong:

- Check not just the `Error` returned by `output()` (which merely tells
  us whether we spawned the clang process successfully) but also the
  `status` field of `std::process::Output` (i.e. the process's exit
  code).

- Forward Clang's stderr to autocxx's stderr.

I ran into an error where the version of Clang on the path did not
understand a certain flag that I was trying to pass (and which is
understood by the Clang I'm setting with `CLANG_PATH`), but I wasn't
getting an error message, which made it harder to diagnose what was
going on.