Use simpler code for 'extern C++' types
Previously, when informing cxx about a type, we always used
  type X = super::bindgen::X;

That was even the case for opaque (non-POD) types.
It results in simpler bindings if we instead tell cxx:
  type X;

That's what this change does.

It works in all circumstances but one. The one circumstance
is nested classes, where cxx is unable to do the correct
C++ codegen. For those, we stick to the old system, so this
unfortunately isn't quite as simple as I'd like.