Adding PinMut trait.
cxx requires that a reference be pinned in order to be able
to call non-const C++ methods on it.

As such, the AsMut trait is not useful to return a reference
on which methods can be called. It turns out to be useful
in client code to have a consistent pattern to return a pinned
mutable reference. Here it is.