Fix variable-length ints in structs.
Previously we regarded int32 as plain-old-data, but not
int. With this change, we consider not only the main name
of the type (e.g. autocxx::c_int) but also any extra names
that the type has (e.g. std::os::raw::int) which is necessary
for the POD checker to identify that these fields are safely POD.

Fixes #667.