Detect opaque types explicitly.
There are some types which bindgen cannot understand. It instead
uses an array in their stead. Previously, autocxx did not understand
arrays and therefore these types were rejected by autocxx, and
substituted for the correct error message.

With #666, we're trying to support actual arrays in autocxx,
and thus we need to disambiguate real arrays from these opaque types.
This change does just that. It relies on autocxx-bindgen emitting
additional annotations.