Support stack allocating C++ objects.
Non trivial C++ objects can now be allocated on the Rust stack using 'moveit':

    moveit! {
      let mut obj = ffi::SomeNonTrivialObject::new();
    }
    obj.do_stuff();

This is just the first step of integrating moveit with autocxx; we also want
to allow transfer of return values out of functions onto the Rust stack and
allow objects to be passed by copy constructor or move constructor into functions.

In general, it's therefore best to stick with the prior behavior where everything
owned by Rust is owned within a cxx::UniquePtr. For now.

Nevertheless this might already be useful for some use cases where high
performance is required.