For non-POD value parameters, take a reference.
This is normal in the moveit world. Work towards #379.
This is because, for value parameters (as opposed to rvalue references)
C++ copies parameters into the function, of course. This means that
it actually (logically) calls the copy constructor. It doesn't consume
the parameter.

This is a major breaking change and we need to think about whether it's
a good idea.