Break circular dependency in casts.
Previously due to a thinko, the function API which did a cast from A->B
would depend on the type A, and the type A would depend on the cast function
A->B. This later caused problems for #379 but might have caused problems
in any case meanwhile.

This removes the latter dependency relationship.