Cease to require bases to be allowlisted.
Previously we required base classes to be on the allowlist,
on the assumption that we were otherwise unable to see functions
belonging to base classes and thus determine whether they
were abstract.

This was possibly a wrong assumption - it looks like bindgen
fully generates base classes for anything on the allowlist.
Let's test that.

Relates to #774.