Always RenameInOutputMod instead of #[rust_name]
Previously we generated slightly simpler output by using
`#[rust_name]` attributes on the cxx::bridge mod. With this change,
we never do that and instead always perform any such renames
with a `use` statement in the output set of mods.

This simplifies our code, while making output more verbose;
it is a prerequisite for a subsequent step towards solving #486.