Try passing uncopyable types by value
It results in generated Rust code that doesn't compile. Not sure if
there's an easy fix, or if this would make more sense as copies of these
tests which are skipped. Also not sure if there's an existing bug that
is intended to include this one, or of it should be a new bug.