Introduce extra phase for special members.
Previously, we mutated an API to fill in a field to indicate
what constructors/special members we'd found.

With this commit, we now add an extra analysis phase where
we add in this field. This eliminates at compile-time any risk
that we'll rely on this field before it is properly set.

This is a little bit more verbose, and potentially marginally
slower, but reduces risk of error for future maintainers.