Use miette for nicer syntax errors.
autocxx code generation is used to parse source code before the Rust compiler
even gets its mitts on it. rustc will never be invoked if autocxx code
generation fails, and that means, for some types of syntax error, the only
messages we'll get are those from autocxx.

Previously autocxx error messages were not useful and did not include details
of the span of affected code. With this PR, we now use thiserror and miette to
format errors more nicely.

This is a breaking change in a couple of ways:
* We have removed the long-deprecated `expect_build()` function
* Error types returned are now a bit different

Fixes #977.