Ensure subclasses gain destructors from supers.
Previously we didn't populate a destructor (Drop) trait for a synthesized
subclass even if the superclass had a destructor. We should have done.

The general problem of derived classes being unaware of their superclass
destructors is #988. The present fix only applies to synthesized subclasses.

As well as being incorrect, this also prevented subclasses gaining a
make_unique method.