Instead return impl New<Output=T>.
This allows callers to put any value on the stack, or in a Box, instead of into a UniquePtr.
A Box has the advantage that it's never NULL, so subsequent use is a little easier.

This is a major breaking change. Callers will now need to use
  .within_unique_ptr()
or
  .within_box()

after every function call they make which returns a non-POD type.

Fixes a major part of #379.