Interpret 'use' typedefs.
Bindgen converts typedefs to two different things:
* type A = B;
* use B as A;

Previously we passed on the second type of typedef into the output mod, but we
made no attempt semantically to understand them. So if 'B' was an enum or POD
type, our typedef analysis wouldn't spot this. Worse, such typedefs were not
exposed to cxx, and therefore we sometimes got 'invalid type' errors if such
types were used in APIs.