Introduce 'instantiable!' directive.
When autocxx encountered a typedef to something that it couldn't comprehend, it
treated that as an 'opaque typedef' such that people could still use it in
limited contexts. One of those contexts was inside a UniquePtr or CxxVector;
unfortunately that proved to be too lenient. Sometimes the referents of such
typedefs would themselves include forward declarations, which were of course
invisible to autocxx since the whole point is that we couldn't see inside the
definition of these typedefs.

So, we now treat such typedefs as non-instantiable by default. This renders
them significantly less useful, so there's now an "instantiable!" directive
that can be used to revert to the prior behavior.

Fixes #1065.