No anon NS types; do not concretize bad templates.
This commit has two fixes (which started off related but then ceased to be):

* If we ignore a template for some reason, we now no longer generate
  a concrete type for usages thereof.
  This may be more conservative than we need. It may be that some reasons
  why templates are ignored do not apply to generated concrete types.
  We may wish to relax this in future. The specific case that triggered
  it was an inner type a<b>::c, where we were unable to generate C++ to
  apply <b> to a instead of c because that information was not present
  in bindgen.
* We do not generate methods for things in anonymous namespaces, because
  we can't refer to them correctly in C++ code.

This latter fix is both too aggressive and too weak. It's too weak because
it just prevents us making methods on these things; if the types are used
as function parameters or return values we will still mis-generate code
containing "_bindgen_mod_XYZ" in C++.

On the other hand it might be that the better fix is to go ahead and
generate C++ code for these things, just stripping out the _bindgen_mod_XYZ
bits.