Fix and test constructor wrappers in multiple include_cpps
The naming convention that just used the C++ function name directly
doesn't work with multiple include_cpps in the same binary. The test
case has them in the same rust file, which could be solved other ways,
but ultimately we may link the generated C++ code from multiple Rust
files which means the names need to be globally unique.