Consider that enums are always movable/destroyable.
The other root cause behind bug #1170 was that C++ enums did not
get move constructors nor destructors generated (they don't need them!)
- but their absence was being used as evidence that any outer type
containing such a field was also not movable.

For example,

  struct C {
    enum D {} some_enum_field;
  };

C would not be considered movable, and therefore could not exist
in a C++ vector.

This change arranges to ensure that enums are always considered
movable or destroyable in this analysis.