Fix use of extern_rust_type in Box<T> in wrappers.
Sometimes autocxx generates wrapper functions on the Rust side, depending on
the needs of some of the arguments and return values. If these wrapper
functions took as a parameter Box<T> where T was an extern_rust_type, then this
failed because T was specified as root::T yet T was not in the root. This
change fixes that problem by ensuring T always is at the root.