Newtype wrappers for some names.
This is an internal change which begins to dig us out of the mess of naming we
have inside autocxx. We deal with all sorts of different kinds of identifiers -
for Rust, for cxx, for C++, from bindgen - and convert between them in a fairly
ad-hoc basis. It's all too fragile to be able to tackle changes like #1371.

Fortunately, Rust makes it easy to solve these sorts of messes using newtype
wrappers. Unfortunately, I've tried that in the past and it's got to be a
huge muddle.

Here's another attempt at carving off part of the space. This introduces
such wrappers for two types of name:
* The [bindgen_original_name] identifiers discovered in bindgen
  attributes;
* The final name to be used for C++ function calls.

There are various FIXMEs added in locations where we're converting to or from
other types of name in questionable ways.

A bit of #520.