Update CppPeerConstructor documentation
These docstrings still refer to ::make_unique(), which has been
deprecated since v0.19.1 and entirely removed since v0.22.0.
Additionally, the top-level documentation said to call it on the
superclass, which as far as I can tell is not correct (it should be
called on the peer object instead, as the make_peer() docstring states).