Synthesize subclass constructor when superclass destructor is protected
Currently, we only synthesize a default constructor when the superclass
destructor is public. This is to comply with 11.4.5.2 of the C++
standard[1], which states that a default constructor is deleted if "any
potentially constructed subobject [e.g. the virtual base class] has a
type with a destructor that is deleted or inaccessible from the
defaulted default constructor".

We correctly exclude private base class destructors, which are not
accessible from the subclass's constructor. But we also exclude
protected base class destructors, which are! Fix that by matching
protected superclass destructors too.

[1] https://github.com/cplusplus/draft/releases/tag/n4917