Omit unique_ptr upcasts when superclass has inaccessible destructor
Currently, we emit as_<superclass>_unique_ptr helpers for all Rust
subclasses. But this breaks subclasses derived from classes with private
or protected destructors, since unique_ptr needs to be able to destruct
the object it references. Fix that by omitting those helpers when we
know the superclass destructor is inaccessible.

The core changes are in codegen_rs and codegen_cpp; everything else is
just plumbing to get information about superclass destructor visibility
where it needs to be.